Move cleanups fixture to common location
The bootstrap tests don't use the old location
