+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from typing import TYPE_CHECKING
 
@@ -7,6 +8,8 @@ from _pytest.config.argparsing import Parser
 
 if TYPE_CHECKING:
     from localstack.testing.snapshots import SnapshotSession
+
+LOG = logging.getLogger(__name__)
 
 os.environ["LOCALSTACK_INTERNAL_TEST_RUN"] = "1"
 
@@ -136,3 +139,16 @@ def secondary_aws_client(aws_client_factory):
     from localstack.testing.aws.util import secondary_testing_aws_client
 
     return secondary_testing_aws_client(aws_client_factory)
+
+
+@pytest.fixture
+def cleanups():
+    cleanup_fns = []
+
+    yield cleanup_fns
+
+    for cleanup_callback in cleanup_fns[::-1]:
+        try:
+            cleanup_callback()
+        except Exception as e:
+            LOG.warning("Failed to execute cleanup", exc_info=e)
